+++ test/integration/client-middleware-cache.js
@@ -5,10 +5,10 @@ import cacheManager from "cache-manager";
 import jwt from "jwt-simple";
 import _ from "lodash";
 
-import Middleware from "../src/middleware/client";
-import { Cache } from "../src/infra";
+import Middleware from "../../src/middleware/client";
+import { Cache } from "../../src/infra";
 
-import HullStub from "./support/hull-stub";
+import HullStub from "../unit/support/hull-stub";
 
 let reqStub;