+++ test/unit/client-middleware.js
@@ -4,7 +4,7 @@ import sinon from "sinon";
 import Promise from "bluebird";
 import _ from "lodash";
 
-import Middleware from "../src/middleware/client";
+import Middleware from "../../src/middleware/client";
 import HullStub from "./support/hull-stub";
 
 describe("Client Middleware", () => {